feat: Show bounty prize on cards
