+++ frontend/src/app/explore/page.tsx
@@ -19,7 +19,8 @@ import { AppBounty } from "../../model/state";
 import { BountyStatusBadgeGroup } from "../../components/bountyStatus";
 import { HasConnectedAccount } from "../../components/hasConnectedAccount";
 import { useBlockTimestamp } from "../../hooks/block";
-import { getBountyStatus } from "../../utils/bounty";
+import { getBountyStatus, getBountyTotalPrize } from "../../utils/bounty";
+import { useErc20Metadata, formatErc20Amount } from "../../utils/erc20";
 
 const Bounty: FC<{
     index: number;
@@ -27,6 +28,9 @@ const Bounty: FC<{
     blockTimestamp: bigint;
 }> = ({ index, bounty, blockTimestamp }) => {
     const bountyStatus = getBountyStatus(bounty, blockTimestamp);
+    const totalPrize = getBountyTotalPrize(bounty);
+    const { token } = bounty;
+    const erc20Metadata = useErc20Metadata(token);
     return (
         <Anchor href={"/bounty/" + index} underline="never">
             <Card h="100%">
@@ -41,12 +45,24 @@ const Bounty: FC<{
                     />
                 </Card.Section>
                 <Box>
-                    <Group my="sm">
-                        <Text truncate="end" fw={700} size="lg">
-                            {bounty.name}
+                    <Stack>
+                        <Group my="sm">
+                            <Text truncate="end" fw={700} size="lg">
+                                {bounty.name}
+                            </Text>
+                            <BountyStatusBadgeGroup
+                                bountyStatus={bountyStatus}
+                            />
+                        </Group>
+                        <Text>
+                            Total Prize:{" "}
+                            {formatErc20Amount(
+                                token,
+                                totalPrize,
+                                erc20Metadata,
+                            )}
                         </Text>
-                        <BountyStatusBadgeGroup bountyStatus={bountyStatus} />
-                    </Group>
+                    </Stack>
                     <Text truncate="end" size="xs" c="dimmend">
                         {bounty.description}
                     </Text>
